requirements.txt
Scrapy==1.5.0
google-cloud-datastore==1.6.0
